dglib_chatbot/pyproject.toml
[project]
name = "dglib-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "DaehyungKim",email = "daehyung0330@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "apscheduler (>=3.11.0,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
start = "dglib_chatbot.app:main"